% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osteoblast_tss_chr1}
\alias{osteoblast_tss_chr1}
\title{Osteoblast tss-level data for chr1}
\format{
SummarizedExperiment
}
\usage{
data(osteoblast_tss_chr1)
}
\description{
Subset of the osteoblast dataset, summarized to TSS level
(50 bp resolution).
}
\references{
The osteoblast differentiation experiment is described here:

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE54461}

The quantification steps are provided here:

\url{https://github.com/mikelove/osteoblast-quant}
}
\keyword{datasets}
